# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/posts": {
      "post": {
        "operationId": "PostsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Related entity not found"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Create a new post",
        "tags": [
          "Posts"
        ]
      }
    },
    "/posts/bulk": {
      "post": {
        "operationId": "PostsController_bulkCreate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkCreatePostsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Posts created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Bulk create posts",
        "tags": [
          "Posts"
        ]
      }
    },
    "/posts/bulk/csv": {
      "post": {
        "operationId": "PostsController_bulkCreateFromCsv",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "CSV with headers: contentAr,contentEn?,scheduleAt,authorId,teamId?,socialAccountIds?(comma-separated),campaignId?,templateId?,status?,isRecurring?"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Posts created from CSV",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Bulk create posts from CSV upload",
        "tags": [
          "Posts"
        ]
      }
    },
    "/posts/recurring": {
      "post": {
        "operationId": "PostsController_createRecurring",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecurringPostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Recurring post created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Create a recurring post",
        "tags": [
          "Posts"
        ]
      }
    },
    "/posts/draft": {
      "post": {
        "operationId": "PostsController_createDraft",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftPostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Draft post created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Create a draft post",
        "tags": [
          "Posts"
        ]
      }
    },
    "/posts/{id}": {
      "put": {
        "operationId": "PostsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Post not found"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Update an existing post",
        "tags": [
          "Posts"
        ]
      },
      "delete": {
        "operationId": "PostsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post marked as deleted"
          },
          "404": {
            "description": "Post not found"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Delete a post (soft delete)",
        "tags": [
          "Posts"
        ]
      }
    },
    "/posts/{id}/reschedule": {
      "put": {
        "operationId": "PostsController_reschedule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReschedulePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post rescheduled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Reschedule a post",
        "tags": [
          "Posts"
        ]
      }
    },
    "/posts/{id}/publish": {
      "post": {
        "operationId": "PostsController_publishNow",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Publish job enqueued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Publish a post immediately (enqueue job)",
        "tags": [
          "Posts"
        ]
      }
    },
    "/posts/{id}/status": {
      "get": {
        "operationId": "PostsController_statusGet",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Status fetched",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostStatusResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Get post status",
        "tags": [
          "Posts"
        ]
      }
    },
...
    "/analytics/posts/{postId}": {
      "get": {
        "operationId": "AnalyticsController_getPostInsights",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostInsightsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Post Insights",
        "tags": [
          "Analytics"
        ]
      }
    },
...
  },
  "info": {
    "title": "YSHAI API",
    "description": "Generated for Orval client",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken"
      }
    },
    "schemas": {
...
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "contentAr": {
            "type": "string",
            "description": "Arabic content of the post",
            "example": "هذا محتوى المنشور باللغة العربية"
          },
          "contentEn": {
            "type": "string",
            "description": "English content of the post (optional)",
            "example": "This is the post content in English"
          },
          "scheduleAt": {
            "type": "string",
            "description": "Scheduled date and time for the post (ISO format)",
            "example": "2025-04-05T10:00:00Z",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "description": "Status of the post",
            "enum": [
              "draft",
              "scheduled",
              "published",
              "failed",
              "pending_approval"
            ],
            "default": "draft"
          },
          "isRecurring": {
            "type": "boolean",
            "description": "Whether the post is part of a recurring schedule",
            "default": false
          },
          "authorId": {
            "type": "number",
            "description": "ID of the author (User)",
            "example": 1
          },
          "teamId": {
            "type": "number",
            "description": "ID of the team (optional)",
            "example": 5
          },
          "socialAccountIds": {
            "description": "IDs of the social accounts to post to (optional, multi-target)",
            "example": [
              12,
              15
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "campaignId": {
            "type": "number",
            "description": "ID of the associated campaign (optional)",
            "example": 7
          },
          "templateId": {
            "type": "number",
            "description": "ID of the template used (optional)",
            "example": 3
          }
        },
        "required": [
          "contentAr",
          "scheduleAt",
          "authorId"
        ]
      },
      "PostResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "contentAr": {
            "type": "string"
          },
          "contentEn": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "scheduled",
              "published",
              "failed",
              "pending_approval"
            ]
          },
          "isRecurring": {
            "type": "boolean"
          },
          "publishedAt": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "scheduleAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "contentAr",
          "status",
          "isRecurring",
          "createdAt",
          "updatedAt",
          "scheduleAt"
        ]
      },
      "BulkCreatePostsDto": {
        "type": "object",
        "properties": {
          "posts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreatePostDto"
            }
          }
        },
        "required": [
          "posts"
        ]
      },
      "RecurringPostDto": {
        "type": "object",
        "properties": {
          "contentAr": {
            "type": "string",
            "description": "Arabic content"
          },
          "contentEn": {
            "type": "string",
            "description": "English content"
          },
          "scheduleAt": {
            "type": "string",
            "description": "First scheduled time (ISO)",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "description": "Status",
            "enum": [
              "draft",
              "scheduled",
              "published",
              "failed",
              "pending_approval"
            ]
          },
          "isRecurring": {
            "type": "boolean",
            "description": "Force recurring",
            "default": true
          },
          "authorId": {
            "type": "number",
            "description": "Author ID"
          },
          "teamId": {
            "type": "number",
            "description": "Team ID"
          },
          "socialAccountIds": {
            "description": "Social account IDs (multi-target)",
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "campaignId": {
            "type": "number",
            "description": "Campaign ID"
          },
          "templateId": {
            "type": "number",
            "description": "Template ID"
          }
        },
        "required": [
          "contentAr",
          "scheduleAt",
          "isRecurring",
          "authorId"
        ]
      },
      "DraftPostDto": {
        "type": "object",
        "properties": {
          "contentAr": {
            "type": "string",
            "description": "Arabic content"
          },
          "contentEn": {
            "type": "string",
            "description": "English content"
          },
          "scheduleAt": {
            "type": "string",
            "description": "When you plan to publish this draft",
            "format": "date-time"
          },
          "authorId": {
            "type": "number",
            "description": "Author ID"
          },
          "teamId": {
            "type": "number",
            "description": "Team ID"
          },
          "socialAccountIds": {
            "description": "Social account IDs (multi-target)",
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "campaignId": {
            "type": "number",
            "description": "Campaign ID"
          },
          "templateId": {
            "type": "number",
            "description": "Template ID"
          }
        },
        "required": [
          "contentAr",
          "scheduleAt",
          "authorId"
        ]
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {
          "contentAr": {
            "type": "string",
            "description": "Arabic content",
            "example": "محتوى محدث"
          },
          "contentEn": {
            "type": "string",
            "description": "English content",
            "example": "Updated content"
          },
          "scheduleAt": {
            "type": "string",
            "description": "Scheduled date/time",
            "example": "2025-04-06T14:30:00Z",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "description": "Status",
            "enum": [
              "draft",
              "scheduled",
              "published",
              "failed",
              "pending_approval"
            ]
          },
          "isRecurring": {
            "type": "boolean",
            "description": "Is recurring?"
          },
          "teamId": {
            "type": "number",
            "description": "Team ID",
            "example": 5
          },
          "socialAccountIds": {
            "description": "Social account IDs (multi-target)",
            "example": [
              12,
              15
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "campaignId": {
            "type": "number",
            "description": "Campaign ID",
            "example": 7
          },
          "templateId": {
            "type": "number",
            "description": "Template ID",
            "example": 3
          }
        }
      },
      "ReschedulePostDto": {
        "type": "object",
        "properties": {
          "scheduleAt": {
            "type": "string",
            "description": "New scheduled date/time",
            "format": "date-time"
          }
        },
        "required": [
          "scheduleAt"
        ]
      },
      "PostStatusResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "scheduled",
              "published",
              "failed",
              "pending_approval"
            ]
          },
          "scheduledAt": {
            "format": "date-time",
            "type": "string"
          },
          "publishedAt": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "status",
          "scheduledAt"
        ]
      },
...
      "PostInsightsDto": {
        "type": "object",
        "properties": {
          "postId": {
            "type": "number"
          },
          "impressions": {
            "type": "number"
          },
          "clicks": {
            "type": "number"
          },
          "likes": {
            "type": "number"
          },
          "comments": {
            "type": "number"
          },
          "shares": {
            "type": "number"
          },
          "fetchedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "postId",
          "impressions",
          "clicks",
          "likes",
          "comments",
          "shares",
          "fetchedAt"
        ]
      },
...
    }
  }
}